Improve `extract_some_wakers()` and `wake_all()` from 30 ns to under 7 ns.
Unfortunately, Rust's optimizer does not understand that returning a
partially-filled ArrayVec on the stack does not require copying every
element, and it generates a significant number of vector instructions
to move ExtractedWakers around on the stack.

This commit makes a mildly unpleasant change to the API that generates
substantially better code and will fix a bottleneck in batch-channel.